inline convertHost function
