+++ hyperbahn-client.js
@@ -29,7 +29,6 @@ var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 var safeJsonParse = require('safe-json-parse/tuple');
 var path = require('path');
-var hyperbahnUtils = require('./hyperbahn/utils');
 
 var Reporter = require('./tcollector/reporter.js');
 var TChannelJSON = require('./as/json.js');
@@ -518,7 +517,7 @@ function discover(opts, cb) {
 
         hosts = [];
         for (var i = 0; i < res.body.peers.length; i++) {
-            hosts.push(hyperbahnUtils.convertHost(res.body.peers[i]));
+            hosts.push(convertHost(res.body.peers[i]));
         }
 
         cb(null, hosts);
@@ -549,4 +548,13 @@ function safeSyncRead(filePath) {
     }
 
     return [error, fileContents];
+}
+
+function convertHost(host) {
+    var res = '';
+    res += ((host.ip.ipv4 >>> 24) & 0xff) + '.';
+    res += ((host.ip.ipv4 >>> 16) & 0xff) + '.';
+    res += ((host.ip.ipv4 >>> 8) & 0xff) + '.';
+    res += (host.ip.ipv4 & 0xff);
+    return res + ':' + host.port;
 }
